requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
PyYAML
asgiref
sqlparse
python-dotenv
django-mongodb-backend
pymongo[srv]
